+++ changelog.md
@@ -1,4 +1,10 @@
 # Changelog
+## 1.0.3
+- Fixed reload button's layout (tysm erymanthus), updated the sprite
+- Fixed compatibility with Cvolton's CompactLists (tysm erymanthus)
+- Fixed compatibility with Global List Integrated
+- Fixed (not really) compatibility with Integrated Demonlist
+- Enhanced inner code structure
 ## 1.0.2
 - removed try-catch block from data loading function
 ## 1.0.1